+++ framework/domain/src/main/java/is/codion/framework/domain/entity/Entity.java
@@ -499,7 +499,7 @@ public interface Entity extends Comparable<Entity> {
 		/**
 		 * @return the entity type
 		 */
-		EntityType entityType();
+		EntityType type();
 
 		/**
 		 * @return the entity definition
@@ -601,7 +601,7 @@ public interface Entity extends Comparable<Entity> {
 		 */
 		static LinkedHashMap<EntityType, List<Key>> groupByType(Collection<Key> keys) {
 			return requireNonNull(keys).stream()
-							.collect(groupingBy(Key::entityType, LinkedHashMap::new, toList()));
+							.collect(groupingBy(Key::type, LinkedHashMap::new, toList()));
 		}
 
 		/**